+++ tests/image_methods_test.py
@@ -2372,6 +2372,16 @@ def test_trim_fuzz(fx_asset):
         assert fuzz_y < trim_y
 
 
+def test_trim_percent_background(fx_asset):
+    # TODO - Find a better test image to demonstrate trim ranges.
+    with Image(filename=str(fx_asset.join('horizon_sunset_border2.jpg'))) as img:
+        img.trim(fuzz=0.0, percent_background=0.0, background_color='black')
+        assert img.size == (400, 298)
+    with Image(filename=str(fx_asset.join('horizon_sunset_border2.jpg'))) as img:
+        img.trim(fuzz=0.0, percent_background=0.5, background_color='black')
+        assert img.size == (472, 344)
+
+
 def test_trim_reset_coords():
     with Image(filename='logo:') as img:
         page = img.page